feat: prepare for packagization of svg libs
