+++ package.json
@@ -44,7 +44,7 @@
     "gatsby-plugin-remove-serviceworker": "^1.0.0",
     "gatsby-theme-carbon": "^1.20.0",
     "html-loader": "^0.5.5",
-    "lodash": "^4.17.15",
+    "lodash-es": "^4.17.15",
     "markdown-it": "^9.0.1",
     "markdown-loader": "^5.0.0",
     "prismjs": "^1.17.1",
