--- plugins/gatsby-plugin-carbon-icon-library/index.js
@@ -1,1 +1,0 @@
-// no-op
