--- plugins/gatsby-plugin-carbon-icon-library/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "gatsby-plugin-carbon-icon-library"
-}
