+++ plugins/gatsby-theme-carbon-svgs/components/IconLibrary/index.js
