+++ plugins/gatsby-theme-carbon-svgs/components/PictogramLibrary/PictogramCategory.js
