+++ plugins/gatsby-theme-carbon-svgs/components/PictogramLibrary/PictogramLibrary.js
@@ -1,19 +1,19 @@
-/* eslint-disable no-restricted-globals */
 /* eslint-disable no-debugger */
 import React, { useEffect, useState } from 'react';
-import { pickBy, groupBy, debounce } from 'lodash';
-import * as iconsReact from '@carbon/icons-react';
 
-import iconMetaData from './iconMetaData';
-import { svgPage, svgLibrary } from '../shared/SvgLibrary.module.scss';
+import { groupBy, debounce } from 'lodash-es';
+import * as pictogramsReact from '@carbon/pictograms-react';
 
 import FilterRow from '../shared/FilterRow';
-import IconCategory from './IconCategory';
+import pictogramMetaData from './pictogramMetaData';
+import { svgPage, svgLibrary } from '../shared/SvgLibrary.module.scss';
+
+import PictogramCategory from './PictogramCategory';
 import NoResult from '../shared/NoResult';
 
 const IconLibrary = () => {
-  const [iconComponents, setIconComponents] = useState([]);
-  const [selectedCategory, setSelectedCategory] = useState('All icons');
+  const [pictogramComponents, setPictogramComponents] = useState([]);
+  const [selectedCategory, setSelectedCategory] = useState('All pictograms');
   const [searchInputValue, setSearchInputValue] = useState('');
   const [categoryList, setCategoryList] = useState([]);
   const [categoriesLoaded, setCategoriesLoaded] = useState(false);
@@ -21,35 +21,23 @@ const IconLibrary = () => {
   const debouncedSetSearchInputValue = debounce(setSearchInputValue, 200);
 
   useEffect(() => {
-    const iconComponentList = pickBy(
-      iconsReact,
-      (val, key) => key.slice(-2) === '32'
-    );
-
-    const iconArray = Object.keys(iconMetaData).map(icon => ({
-      ...iconMetaData[icon],
+    const iconArray = Object.keys(pictogramMetaData).map(icon => ({
+      ...pictogramMetaData[icon],
       // If the icon is unprefixed and starts with a number, add an underscore
-      Component:
-        iconComponentList[isNaN(icon[0]) ? `${icon}32` : `_${icon}32`] ||
-        iconComponentList[`WatsonHealth${icon}32`] ||
-        iconComponentList[`Q${icon}32`],
+      Component: pictogramsReact[icon],
     }));
-
-    const filteredIcons = iconArray.filter(({ deprecated }) => !deprecated);
-
     setCategoryList(
-      Object.keys(groupBy(filteredIcons, 'categories[0].name')).sort()
+      Object.keys(groupBy(iconArray, 'categories[0].name')).sort()
     );
     setCategoriesLoaded(true);
-
-    setIconComponents(filteredIcons);
+    setPictogramComponents(iconArray);
   }, []);
 
-  const getFilteredIcons = () => {
+  const getFilteredPictorams = () => {
     if (!searchInputValue) {
-      return iconComponents;
+      return pictogramComponents;
     }
-    return iconComponents.filter(
+    return pictogramComponents.filter(
       // eslint-disable-next-line camelcase
       ({ friendly_name, categories, aliases = [], name }) => {
         const searchValue = searchInputValue.toLowerCase();
@@ -64,23 +52,20 @@ const IconLibrary = () => {
           (categories &&
             categories[0] &&
             categories[0].name.toLowerCase().includes(searchValue)) ||
-          (categories &&
-            categories[0] &&
-            categories[0].subcategory.toLowerCase().includes(searchValue)) ||
           name.toLowerCase().includes(searchValue)
         );
       }
     );
   };
 
-  const filteredIcons = getFilteredIcons();
+  const filteredPictograms = getFilteredPictorams();
 
   const allCategories = Object.entries(
-    groupBy(filteredIcons, 'categories[0].name')
+    groupBy(filteredPictograms, 'categories[0].name')
   );
 
   const filteredCategories =
-    selectedCategory === 'All icons'
+    selectedCategory === 'All pictograms'
       ? allCategories
       : allCategories.filter(([category]) => category === selectedCategory);
 
@@ -89,6 +74,7 @@ const IconLibrary = () => {
   return (
     <div className={svgPage}>
       <FilterRow
+        type="pictogram"
         categoryList={categoryList}
         selectedCategory={selectedCategory}
         onSearchChange={e =>
@@ -100,16 +86,23 @@ const IconLibrary = () => {
       />
       {shouldShowNoResult ? (
         <NoResult
+          type="pictograms"
           selectedCategory={selectedCategory}
           setSelectedCategory={setSelectedCategory}
-          allIconResults={filteredIcons.length}
-          pageName="icon"
-          pageUrl="https://github.com/carbon-design-system/carbon/blob/master/packages/icons/master/ui-icon-master.ai"
+          allIconResults={filteredPictograms.length}
+          pageName={'pictogram'}
+          pageUrl={
+            'https://github.com/carbon-design-system/carbon/blob/master/packages/pictograms/master/pictogram-master.ai'
+          }
         />
       ) : (
         <div className={svgLibrary}>
-          {filteredCategories.map(([category, icons]) => (
-            <IconCategory key={category} category={category} icons={icons} />
+          {filteredCategories.map(([category, pictograms]) => (
+            <PictogramCategory
+              key={category}
+              category={category}
+              pictograms={pictograms}
+            />
           ))}
         </div>
       )}
