+++ plugins/gatsby-theme-carbon-svgs/components/PictogramLibrary/index.js
