+++ plugins/gatsby-theme-carbon-svgs/components/shared/ActionBar.js
