+++ plugins/gatsby-theme-carbon-svgs/components/shared/LibraryProvider.js
