+++ plugins/gatsby-theme-carbon-svgs/components/shared/SvgCard.js
