+++ plugins/gatsby-theme-carbon-svgs/components/shared/SvgLibrary.module.scss
@@ -329,3 +329,16 @@ $svg-card-border: 1px solid $ui-03;
 .no-result :global(.resource-card-group) img {
   width: 32px;
 }
+
+.icon {
+  width: 32px;
+  object-fit: contain;
+  position: relative;
+  display: inline-block;
+  user-select: none;
+  vertical-align: text-bottom;
+}
+
+.icon-inverse {
+  fill: $inverse-01;
+}
