+++ plugins/gatsby-theme-carbon-svgs/components/shared/useSticky.js
@@ -1,6 +1,6 @@
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useState, useRef } from 'react';
 
-export const useSticky = () => {
+export default () => {
   const stickyRef = useRef(null);
   const [sticky, setSticky] = useState(false);
   const eventsToBind =
@@ -41,24 +41,4 @@ export const useSticky = () => {
   }, [stickyRef, sticky, eventsToBind]);
 
   return [stickyRef, sticky];
-};
-
-export const useOnClickOutside = (ref, handler) => {
-  useEffect(() => {
-    const listener = event => {
-      // Do nothing if clicking ref's element or descendents
-      if (!ref.current || ref.current.contains(event.target)) {
-        return;
-      }
-
-      handler(event);
-    };
-    document.addEventListener('mousedown', listener);
-    document.addEventListener('touchstart', listener);
-
-    return () => {
-      document.removeEventListener('mousedown', listener);
-      document.removeEventListener('touchstart', listener);
-    };
-  }, [ref, handler]);
 };
