+++ plugins/gatsby-theme-carbon-svgs/gatsby-node.js
