+++ plugins/gatsby-theme-carbon-svgs/index.js
@@ -1,0 +1,3 @@
+// no-op
+export { default as PictogramLibrary } from './components/PictogramLibrary';
+export { default as IconLibrary } from './components/IconLibrary';
