+++ plugins/gatsby-theme-carbon-svgs/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "gatsby-theme-carbon-svgs",
+  "dependencies": {
+    "classnames": "^2.2.6",
+    "copy-to-clipboard": "^3.2.1",
+    "lodash-es": "^4.17.15",
+    "change-case": "^4.1.1",
+    "@carbon/icons": "^10.8.2",
+    "@carbon/icons-react": "^10.8.2",
+    "@carbon/pictograms": "^10.8.2",
+    "@carbon/pictograms-react": "^10.8.2",
+    "carbon-components": "^10.9.3",
+    "carbon-components-react": "^7.9.3"
+  },
+  "peerDependencies": {
+    "gatsby-theme-carbon": "^1.20.1",
+    "react": "^16.12.0",
+    "gatsby": "^2.19.17"
+  }
+}
